requirements.txt
Django==4.2
django-crispy-forms==2.2
crispy-bootstrap4==2024.1
# git+https://github.com/titipata/pubmed_parser.git
pubmed-parser
# Scholar Scraper Slightly Modified
#./deployment-infra/vendor/pubmed_parser
#git+https://github.com/ethandrower/scholarly.git
scholarly==1.7.11
httpx==0.27.0
biopython==1.85
Pillow==9.2.0
whitenoise==6.2.0
pandas==2.0.0
django-storages==1.12.3
fuzzywuzzy==0.18.0
celery==5.4.0
# redis==4.3.4
dj-database-url==0.5.0
psycopg2==2.9.9
gunicorn==20.1.0
django-debug-toolbar==4.0.0
boto3==1.37.23
django-floppyforms==1.9.0
rispy==0.7.1
python-dotenv===1.0.0
python-docx==0.8.11
loguru==0.6.0
django-mailgun-provider==0.2.3
python-Levenshtein==0.27.1
bs4==0.0.1
selenium==4.11.2
webdriver-manager==3.8.2

docxtpl==0.16.0
django-s3direct==2.0.0
djangorestframework==3.15.2

# develop dependencies
flake8==5.0.4
black==21.12b0
pytest-django==4.5.2
django-extensions==3.1.5
openpyxl==3.1.0
xlrd==2.0.1
XlsxWriter==3.1.0
inflect==6.0.4 
pydantic==2.10.6
PyMuPDFb==1.23.3 
pymupdf==1.25.5

contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.1
packaging==23.2
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
django-activity-stream==2
django-cors-headers==4.6.0

# djangop channels
attrs==25.3.0
autobahn==24.4.2 
automat==25.4.16 
channels==4.2.2 
constantly==23.10.4 
daphne==4.1.2 
hyperlink==21.0.0 
incremental==24.7.2 
pyasn1==0.6.1
pyasn1-modules==0.4.2 
service-identity==24.2.0 
twisted==24.11.0 
txaio==23.1.1 
zope-interface==7.2
async-timeout==5.0.1 
channels_redis==4.2.1 
msgpack==1.1.0 
redis==5.2.1

# AI 
git+https://x-token-auth:ATCTT3xFfGN03gNSI1nF5PU51-OedTJbBCO1V6OC7wng_KfQBkMjlCWwiNcMOfA2ZhtBA__rARVtUsTIZeaYx0ZkXHsWQU98OIN_K1hrAD3x_0Nw3wWcuk_SdmNiW4zewfhUGRKLvOYYw3obOeTKQIKcdjVzKddh_VnhExvJpS9dIsIAfmcvISg=4D3CD778@bitbucket.org/citemed/ai_backend@v1.1.0#egg=citemed_ai